+++ src/javascript/common/champion.js
@@ -18,7 +18,7 @@ const Cashier             = require('./../pages/cashier/cashier');
 const CashierTopUpVirtual = require('./../pages/cashier/top_up_virtual');
 const CashierPaymentMethods = require('./../pages/cashier/payment_methods');
 const CashierPassword       = require('./../pages/cashier/cashier_password');
-const FinancialAssessment   = require('./../pages/user/settings/financial_assessment');
+const FinancialAssessment   = require('./../pages/user/financial_assessment');
 
 const Champion = (function() {
     'use strict';
